move landing Layout from suite structure to landing-page
