+++ packages/landing-page/pages/wallet/start/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import Layout from '@landing-components/Layout';
+import Layout from '@landing-components/StartLayout';
 import { H2, Button, P, Select, Link, variables } from '@trezor/components';
 
 const Wrapper = styled.div`
